Implement regular expression patterns for string generation
When creating a template, for strings that must match regular
expressions in the schema, the pattern is copied into the template. The
generator uses the patterns to generate matching random string.

The generator is not perfect, but should handle handle the regexps found
in the CSAF JSON schema. See the comment for the Sample method in
pkg/fakedoc/regexp.go for details.