Make order of object property generation predictable
In order to make fakedoc output repeatable, the properties of objects
must always be generated in the same order because it affects the state
of the random number generator. In the JSON output the properties are
always ordered alphanumerically because of how JSON serialization of
maps is implemented in encoding/json. To support this, the templates now
represent the property settings as a slice of Property structs, which
includes name, type and probability. This is reflected in the template
file.