Implement uri format for strings by settings a suitable regular expression
An easy way to generate URIs that are at least syntactically valid is to
use a suitable regular expression, since we already have code that can
generate random matches. So, when creating the template, the code now
sets a predefined regular expression for strings with the uri format, if
the schema does not already include a pattern.