Add UniqueItems setting for arrays
The CSAF schema requires unique items for some of the arrays. In most
cases collisions between randomly generated items are very unlikely, but
in the case of vulnerabilities/.../flags it actually occurs occasionally
(e.g. with seed pcg:51:51 before this commit).

We now avoid this with the UniqueItems flag in TmplArray which is set
automatically from the schema when the template is generated. When
generating a random array where UniqueItems is true, we check that items
are unique and attempt to generate a new value if collisions occur. The
number of retries is limited and it can theoretically happen that the
code gives up and exits with an error. I haven't been able to find a
seed for which this happens without modifying the template in a way that
makes this much more likely.

The implementation has quadratic runtime because it performs a linear
search of the partially created array when checking for collisions. This
doesn't seem to be a problem in practice because the arrays are very
short, most of the time.