fakedoc: the template can now replace only parts of the template
fakedoc now always uses a built-in template that is derived from the
CSAF JSON schema and doesn't require the --template argument anymore. If
a template is given, that file is read and the types in it replace the
ones in the built-in template. This way one can influence how the
generator works with very short files that only contain the parts that
have to be different.

README.md and the template documentation have been updated. The latter
includes a section about how this works and an example template that
influences some of the publisher data.