Implement bulk generation of CSAF documents
The fakedoc command has a new option -n for the number of documents to
generate. If that number is greater than 1, the -o option must be used
and is used as a template for the filenames.

This uses text/template as the templating engine, which is a bit too
powerful, but it behaves more nicely than e.g. fmt.Sprintf in case of
errors.