Make schemas with the same underlying type use the same type name
When deriving the template from the schema, some schemas have the same
uderlying type, because the both use  to refer to the same type.
So far, the code correctly used the definition of that shared type, but
used different names when putting it into the template's Types map. With
this change, that name is derived from the shared underlying type.

This leads to more sharing of types which makes it easier to deal with
special handling of e.g. the product_id type which is referenced in many
places.