Implement abandoning of subtrees of the generated tree more robustly
During the construction of the tree for the random CSAF document, it can
happen that a particular branch cannot be fully constructed, because
some constraint cannot be met. The main reason for that so far is that
the maximum depth would be exceeded. In this case the generateNode
method returns an error that indicates this and the generator will try
to recover from this by trying something else, like not generating the
object property where the problem occurred and trying to add different
one.

The implementation of produt IDs will add onother reason for pruning a
branch, so we introce an error value that forms the basis for all
reasons for pruning.

We also now handle these errors in more places, even though those are
currently not places where these error can occur for the built-in
template.