Implement product IDs and references to them
The product tree contains products that are identified with IDs that may
be referenced from elsewhere in the document. Those references must be
generated in such a way that they point to existing products. With this
commit the can do that, although at the moment it doesn't work well
enough yet. In about 2.5% of all attempts, no document will be generated
at all.

The basic approach is to have two new template types, one to generate
IDs and another for places that reference them. During generation
references are only created at all if at least one ID has been generated
and they're initially added with a placeholder. Once the full tree has
been generated, the placeholders are filled with IDs randomly chosen
from all that were generated.

Because this will have to be done with other kinds of IDs, such as group
IDs, the new template types have a namespace parameter is used to
distinguish them.