Add special case for generating arrays of unique references
When generating group_ids we need to be able to generate arrays of with
at least two product_id references all of which have to be different.

With the current approach that doesn't work because the randomArray
method ends up instantiating TmplRef a bunch of times, but all the
reference values generated are equal as they don't have an actual value
yet, and because they're equal randomArray rejects all but the first one
because of the uniqueness requirement and then the array is too short
because it must contain at least two values.

The solution chosen here is to add a special case to randomArray: if the
items are TmplRef and there's a unique constraint we add a placeholde
that represents an array of distinct references which are filled in at
the end like the other references. This new kind of reference is
represented by the same struct type as the other reference placeholder,
but it now has an dditional length field that is used to distinguish the
cases (see the comments for the reference struct).