Add dependency relation between properties of the same object
When generating product groups, we need to generate the product_ids
property before the group_id property. Group IDs are defined by the
product_groups items. If generating the product_ids for a group fails
because e.g. there are no product_ids the entire product group fails and
will not be in the document. If we had already generated a group_id for
the group the generator would have a known group ID and therefore
happily generate references to it elsewere in et document, but the
document would not actually define it leading to dangling references.

The dependency between properties allows us to indicate that group_id
depends on product_ids and therefore the latter must be attempted first.