Add special handling for group_ids
The group IDs are now handled basically in the same way as the product
IDs. The built-in template is automatically modified with group ID
specific settings that mostly work like the ones for the product IDs,
with one exception, the group IDs use the dependency mechanism between
properties to make sure that in product_group objects the group ID is
only generated when there are product IDs available.