Implement a better way to deal with group_ids
The depends attribute for properties is a work around for the problem
that fakedoc could generate references to non existing group IDs, which
happened because once an ID was generated it couldn't be removed from
the name space even if the branch for which it was generated had been
abandoned during generation for e.g. lack of product IDs. See 6a957d1
for more details.

As it turns out, there's an easy way of removing IDs generated in
abandoned branches¹: We take a snapshot of the generator's name space
state before attempting to create a branch (basically the entirety of
generateNode), attempt to create the branch, and if it fails because of
an error based on ErrBranchAbandoned we restore the generator's
namespace state to the snapshot.

The key insight for why this works, is that if during the attempt to
generate the branch, any group IDs have been created then any references
that might have been created in that branch because of the existence of
those IDs are also in that branch. So restoring the snapshot removes all
the consequences of ID generation that happened in the attempt.

So, forcing the order in which some properties are generated is not
needed any more. It was a work around that only worked in highly
specific circumstances (it was introduced specifically for product
groups) and is hard to explain to users who want to modify the templates
because it requires somewhat detailed knowlege about how fakedoc works.
The new approach works better and doesn't need to be understood by users
for IDs to work.

¹ branch in the generic tree sense, not necessarily CSAF branches_t