Set upper bound of array sizes based on the limits file
If fakedoc is invoked with a limits file (option -l) the upper bound of
array sizes is now taken from that file. The limits are very high, so to
avoid exceedingly large files, the bounds are multiplied with a size
factor that can be changed with the --size option. By default that
factor is 0.00001. With that value file sizes are usually less than 1
MB. Setting it to e.g. 0.0001 (only 10 times higher) can already result
in files of several hundred MBs.