Improve performance of generating arrays with uniqueItems constraint
The uniqueItems constraint for arrays was implemented inefficiently with
a linear search over the already generated items. This leads to
quadratic run-time behavior. So far this wasn't much of a problem
because arrays tended to be short and using a hash map to keep track of
already generated items wasn't easy because the items might not be
hashable because the can contain slices.

The code now uses a serialization based on the gob module to produce a
hash value so that we can use a hash map to detect collisions. This can
make the generation substantially faster.

Based on a patch by Sascha L. Teichmann <sascha.teichmann@intevation.de>