Handle corner cases of lengths in Generator.book
In particular, if minlength is given in the template and maxlength is
taken from the limit file, maxlength may be less than minlength. In that
case we now make maxlength equal to minlength.

Also, the argument of IntN must be maxlength-minlength+1 so that
maxlength is actually a value that might be generated. This will also
handle the case of maxlength == minlength correctly and without a panic.